requirements.txt
streamlit
selenium
playwright
openai
python-dotenv